Bug fix.
This was the old behavior that got broken with the
last bash code update, which changed this function
into checking if it is an absolute path - as opposed
to whether or not it contains any slashes.

Without this fix you'd find the same executables
with "ls" as with "./ls":

$ which ./ls
/usr/bin/ls

while the correct reply is:
$ which ./ls
/usr/bin/which: no ls in (.)

Having any '/' at all in the search name means that
you DO specify a directory, either absolute, or relative
to the current directory (./...) or relative to your
home direction (~/...) and in all cases the PATH should
no longer be searched.